migrate player think tests from combatant / combative to player ai
