--- engine/test/entities/character/combatant/think.test.ts
@@ -1,74 +1,0 @@
-import { expect } from '@jest/globals';
-import mockMap from '../../../testHelpers/mockMap';
-import { Combatant } from '../../../../js/entities/character/Combatant';
-import Entity from '../../../../js/entities/character/Entity';
-import SentientEntity from '../../../../js/entities/character/SentientEntity';
-
-// https://stackoverflow.com/a/54475733/5450892
-jest.mock('@/engine/js/entities/character', () => require('../../../testHelpers/helpers').createMock);
-jest.mock('@/engine/js/ai/basic',  () => require('../../../testHelpers/helpers').createMock);
-jest.mock('@/engine/js/ai/predator',  () => require('../../../testHelpers/helpers').createMock);
-jest.mock('@/engine/js/mapping/GameMap', () => mockMap);
-jest.mock('@/engine/js/coordinates/point', () => {
-    return {
-        __esModule: true, // this property makes it work
-        default: jest.fn().mockImplementation((x, y) => {})
-    }
-});
-jest.mock('@/engine/js/events', () => {
-    return {
-        __esModule: true, // this property makes it work
-        default: {
-            Subscribe: jest.fn().mockImplementation(() => { }),
-            RaiseEvent: jest.fn().mockImplementation(() => { }),
-            List: new Proxy({}, {
-                get: function(target, name) {
-                    return name;
-                },
-                set: function(target, name, value) {
-                    return true;  // Indicate that the assignment succeeded
-                }
-            })
-        }
-    };
-});
-
-// TODO: these should have nothing to do with combatant / combative once we extract sentient
-
-describe('Combatant.think', () => {
-
-    let combatant: Combatant;
-    let secondEntity: Entity;
-    beforeEach(() => {
-        combatant = new Combatant({
-            ai: null,
-            isPlayer: true
-        });
-        secondEntity = new Entity({});
-    });
-    
-    Entity.prototype.getClosestEntity = jest.fn().mockImplementation((options) => {
-        return secondEntity;
-    });
-
-    it('will super think', () => {
-        const spy = jest.spyOn(SentientEntity.prototype, 'think');
-
-        combatant.think();
-    
-        expect(spy).toHaveBeenCalled();
-    
-        spy.mockRestore();
-    });
-
-    it('if player, sets target to closest entity', () => {
-
-        expect(combatant.target).toBe(undefined);
-
-        combatant.isPlayer = true;
-
-        combatant.think();
-
-        expect(combatant.target).toBe(secondEntity);
-    });
-});
